Spring Data JPA [#505127]

+++ src/main/java/ru/job4j/model/Photo.java
@@ -1,8 +1,8 @@
 package ru.job4j.model;
 
+import jakarta.persistence.*;
 import lombok.*;
 
-import javax.persistence.*;
 import java.util.Objects;
 
 @Entity
@@ -16,7 +16,7 @@ public class Photo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @EqualsAndHashCode.Include
-    private Integer id;
+    private Long id;
 
     private String name;
     private String path;
